+++ docs/.vitepress/components/Demo.vue
@@ -73,7 +73,7 @@ html.dark .demo-preview > *:first-child {
     margin: 0 auto;
 }
 
-.demo-preview > *:not(:first-child, .ipl-expansion-panel__content) {
+.demo-preview > *:not(:first-child, .ipl-expanding-space__content) {
     background-color: var(--vp-c-bg);
     border-radius: 8px;
     padding: 8px;